requirements.txt
asgiref==3.5.2
autopep8==1.6.0
dj-database-url==0.5.0
Django==4.0.6
Django-Basic-CRUD-Generator==0.2
django-on-heroku==1.1.2
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyMySQL==1.0.2
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.2.0
